requirements.txt
asgiref==3.4.1
dj-database-url==1.0.0
Django==3.2.7
django-cors-headers==3.8.0
django-cors-middleware==1.5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
psycopg2==2.9.4
psycopg2-binary==2.9.4
PyJWT==2.3.0
python-dotenv==0.21.0
pytz==2021.1
sqlparse==0.4.1
